simplify Project Generator
